refactor: re-construct service room
